Changed label to int type and implemented multinomialNB

+++ preprocessing.py
@@ -14,8 +14,7 @@ def fetch_dataset():
     df['content'] = df['content'].apply(clean_text)
     df.loc[df.label == "fake", 'label'] = 1
     df.loc[df.label == "real", 'label'] = 0
-    # print(df.info())
-    # Split Train and Test sets
+    df['label'] = df['label'].astype('int')
     X_train, X_test, y_train, y_test = train_test_split(df["content"], df["label"], test_size = .15,
                                                         random_state=0)
     return X_train, X_test, y_train, y_test
@@ -33,8 +32,6 @@ def vectorizer(X_train):
     vectorizer = CountVectorizer(binary=True, stop_words=stopwords.words('english'), 
                              lowercase=True, max_features=5000)
     X_train_fit = vectorizer.fit_transform(X_train)
-    # print(X_train_fit)
-    # print(X_train_fit.toarray())
     return X_train_fit, vectorizer
 
 def preprocessing():
